chore: Restructured folders in /applications. Removed unused files
