+++ src/features/instance/applications/components/ApplicationsSidebar/EmptyApplicationsView.tsx
