+++ src/features/instance/applications/components/ApplicationsSidebar/FileTreeExplorer/filetree.css
