+++ src/features/instance/applications/components/ApplicationsSidebar/FileTreeExplorer/index.tsx
