+++ src/features/instance/applications/index.tsx
@@ -1,27 +1,24 @@
-import { Button } from '@/components/ui/button';
-import { Link } from '@tanstack/react-router';
-import { Edit, FolderPlus } from 'lucide-react';
+import { useInstanceClientIdParams } from '@/config/useInstanceClient';
+import { EditorViewProvider } from '@/features/instance/applications/context/EditorViewProvider';
+import { ApplicationsSidebar } from '@/features/instance/applications/components/ApplicationsSidebar';
+import { getComponentsQueryOptions } from '@/features/instance/operations/queries/getComponents';
+import { useSuspenseQuery } from '@tanstack/react-query';
+import { TextEditorView } from './components/TextEditorView';
 
-export function ApplicationsIndex() {
+export function ApplicationsEditor() {
+	const instanceParams = useInstanceClientIdParams();
+	const { data: getComponentsQueryData } = useSuspenseQuery(getComponentsQueryOptions(instanceParams));
+
 	return (
-		<div className="flex flex-col justify-center gap-4 min-h-[calc(80vh-theme(spacing.20))]">
-			<h1 className="text-2xl font-bold text-center text-white">Applications</h1>
-			<div className="flex flex-col items-center justify-center gap-8 md:flex-row">
-				<Link to="editor" className="w-full max-w-80">
-					<Button className="w-full py-10 text-lg" variant="positiveOutline">
-						<Edit />
-						Edit Applications
-					</Button>
-				</Link>
-				<Link to="new" className="w-full max-w-80">
-					<Button className="w-full py-10 text-lg" variant="positiveOutline">
-						<FolderPlus />
-						Create/Import A New Application
-					</Button>
-				</Link>
+		<EditorViewProvider>
+			<div className="grid grid-cols-1 gap-4 md:grid-cols-12 h-[calc(100vh-theme(spacing.32))]">
+				<section className="h-0 min-h-full col-span-1 overflow-y-scroll text-white md:col-span-4 lg:col-span-3">
+					<ApplicationsSidebar fileTreeQueryData={getComponentsQueryData} />
+				</section>
+				<section className="h-full col-span-1 text-white md:col-span-8 lg:col-span-9">
+					<TextEditorView />
+				</section>
 			</div>
-		</div>
+		</EditorViewProvider>
 	);
 }
-
-
