+++ src/features/instance/applications/routes.ts
@@ -1,7 +1,6 @@
 import { createInstanceLayoutRoute } from '@/features/instance/instanceLayoutRoute';
 import { createRoute } from '@tanstack/react-router';
-import { NewApplications } from '@/features/instance/applications/new';
-import { ApplicationsEditor } from '@/features/instance/applications/editor';
+import { ApplicationsEditor } from '@/features/instance/applications';
 
 export function createApplicationsRoutes(instanceLayoutRoute: ReturnType<typeof createInstanceLayoutRoute>) {
 	const instanceApplicationsIndexRoute = createRoute({
@@ -9,16 +8,8 @@ export function createApplicationsRoutes(instanceLayoutRoute: ReturnType<typeof 
 		path: 'applications',
 		component: ApplicationsEditor,
 	});
-
-	const instanceApplicationsNewRoute = createRoute({
-		getParentRoute: () => instanceLayoutRoute,
-		path: 'applications/new',
-		component: NewApplications,
-	});
-
+	
 	return [
 		instanceApplicationsIndexRoute,
-		instanceApplicationsNewRoute,
-		// instanceApplicationsEditorRoute,
 	];
 }
